+++ packages/auth-utils/src/index.ts
@@ -1,3 +1,3 @@
 export { CogniteAuth } from './authentication';
-export { getFlow, saveFlow } from './utils';
+export { getFlow, saveFlow, removeFlow } from './utils';
 export type { AuthFlow, AuthenticatedUser } from './types';